Final back-end
this is the final back-end job.
This app is now able to:

Create an application that will allow a user to view transaction data from the Ethereum blockchain associated with a specific wallet address W that the user inputs, starting with block B. The application should get information on:

wallets (addresses) and

amounts of ETH associated with transactions made to and from the given wallet W and

show them in a simple human-readable way (ideally, through a web page).

The application should collect and display ALL transaction data starting from the given block B.

The app is also able to:
Given a date in YYYY-MM-DD format, the program should return the exact value of ETH that was available on the given address at YYYY-MM-DD 00:00 UTC time.